# Conflicts:
#	External/Plugins/OutputPanel/PluginUI.cs

+++ PluginCore/ScintillaNet/Configuration/ConfigurationUtility.cs
@@ -1,10 +1,9 @@
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using System.Windows.Forms;
-using System.Collections.Generic;
-using System.Text.RegularExpressions;
 using System.Xml.Serialization;
+using PluginCore;
 using PluginCore.Helpers;
 
 namespace ScintillaNet.Configuration
@@ -18,40 +17,38 @@ namespace ScintillaNet.Configuration
 
         protected virtual byte[] LoadFile(string filename, ConfigFile parent)
         {
-			System.IO.Stream res;
-			byte[] buf;
-			res = OpenFile(filename, parent);
-			if (res != null)
-			{
-				buf = new byte[res.Length];
-				res.Read(buf ,0 ,buf.Length);
-				return buf;
-			}
-			return null;
-		}
+            Stream res;
+            byte[] buf;
+            res = OpenFile(filename, parent);
+            if (res != null)
+            {
+                buf = new byte[res.Length];
+                res.Read(buf ,0 ,buf.Length);
+                return buf;
+            }
+            return null;
+        }
 
         protected virtual Stream OpenFile(string filename, ConfigFile parent)
         {
-			Stream res;
+            Stream res;
             filename = filename.Replace("$(AppDir)", PathHelper.AppDir);
             filename = filename.Replace("$(UserAppDir)", PathHelper.UserAppDir);
             filename = filename.Replace("$(BaseDir)", PathHelper.BaseDir);
             if (File.Exists(filename)) res = new FileStream(filename, FileMode.Open, FileAccess.Read);
-			else res = _assembly.GetManifestResourceStream(String.Format( "{0}.{1}" , _assembly.GetName().Name, filename.Replace("\\" , "." )));
-			if (res == null && parent != null && parent.filename != null)
-			{
-				int p = parent.filename.LastIndexOf('\\');
-				if (p > 0) return OpenFile(String.Format( "{0}\\{1}", parent.filename.Substring(0, p), filename), null);
-			}
-			return res;
-		}
+            else res = _assembly.GetManifestResourceStream(String.Format( "{0}.{1}" , _assembly.GetName().Name, filename.Replace("\\" , "." )));
+            if (res == null && parent != null && parent.filename != null)
+            {
+                int p = parent.filename.LastIndexOf('\\');
+                if (p > 0) return OpenFile(String.Format( "{0}\\{1}", parent.filename.Substring(0, p), filename), null);
+            }
+            return res;
+        }
 
         protected object Deserialize(TextReader reader, Type aType)
         {
-            object local = null;
-            XmlSerializer xmlSerializer = null;
-            xmlSerializer = new XmlSerializer(aType);
-            local = xmlSerializer.Deserialize(reader);
+            XmlSerializer xmlSerializer = XmlSerializer.FromTypes(new[]{aType})[0];
+            object local = xmlSerializer.Deserialize(reader);
             reader.Close();
             return local;
         }
@@ -92,12 +89,15 @@ namespace ScintillaNet.Configuration
                         String original = File.ReadAllText(filename);
                         String overriding = File.ReadAllText(filename + ".override");
                         String tabContent = overriding.Replace("\n", "\n\t\t\t");
-                        Int32 indexStart = original.IndexOf(coloringStart);
-                        Int32 indexEnd = original.IndexOf(coloringEnd);
-                        String replaceTarget = original.Substring(indexStart, indexEnd - indexStart + coloringEnd.Length);
-                        String finalContent = original.Replace(replaceTarget, tabContent);
-                        File.WriteAllText(filename, finalContent);
-                        File.Delete(filename + ".override");
+                        Int32 indexStart = original.IndexOfOrdinal(coloringStart);
+                        Int32 indexEnd = original.IndexOfOrdinal(coloringEnd);
+                        if (indexStart > -1)
+                        {
+                            String replaceTarget = original.Substring(indexStart, indexEnd - indexStart + coloringEnd.Length);
+                            String finalContent = original.Replace(replaceTarget, tabContent);
+                            File.WriteAllText(filename, finalContent);
+                            File.Delete(filename + ".override");
+                        }
                     }
                     catch { /* NO ERRORS... */ }
                 }
@@ -121,12 +121,12 @@ namespace ScintillaNet.Configuration
 
         public virtual object LoadConfiguration(Type configType)
         {
-            return LoadConfiguration( configType, "ScintillaNET.xml", null);
+            return LoadConfiguration(configType, "ScintillaNET.xml", null);
         }
 
         public virtual object LoadConfiguration(Type configType, string filename)
         {
-            return LoadConfiguration( configType, filename, null);
+            return LoadConfiguration(configType, filename, null);
         }
 
         public virtual object LoadConfiguration(string[] files)
